Fix 2.2, 2.3

+++ base-practica1-Qt-2023/remotelink_messages.h
@@ -86,13 +86,12 @@ typedef struct {
 } MESSAGE_OVERSAMPLE_PARAMETER;
 
 typedef struct {
-    uint32_t frecuency;
+    float frecuency;
 } MESSAGE_ADC_AUTO_PARAMETER;
 
 typedef struct {
-    uint32_t frecuency;
+    float frecuency;
 } MESSAGE_ADC_AUTO_FRECUENCY_PARAMETER;
-
 
 typedef struct {
     uint16_t chan[6][16];
